+++ src/test/java/com/beust/jcommander/JCommanderTest.java
@@ -913,6 +913,7 @@ public class JCommanderTest {
     j.usage();
   }
 
+  @Test(expectedExceptions = ParameterException.class, expectedExceptionsMessageRegExp = "Was passed main parameter '' but no main parameter was defined")
   public void tmp() {
     class A {
       @Parameter(names = "-b")
@@ -1079,6 +1080,36 @@ public class JCommanderTest {
     StringBuilder sb = new StringBuilder();
     new JCommander(a).usage(sb);
     Assert.assertTrue(sb.toString().contains("Default: <empty string>"));
+  }
+
+  @Test
+  public void emptyStringShouldBeConsideredAsParameter(){
+    class Arg {
+      @Parameter(description = "parameters")
+      List<String> params;
+    }
+
+    Arg a = new Arg();
+    String [] args = {""};
+
+    new JCommander(a).parse(args);
+    Assert.assertEquals(a.params.size(), 1);
+//    Assert.assertEquals();
+  }
+
+  @Test
+  public void doubleQuotedStringShouldBeConsideredAsParameter(){
+    class Arg {
+      @Parameter(description = "parameters")
+      List<String> params;
+    }
+
+    Arg a = new Arg();
+    String [] args = {"\"\""};
+
+    new JCommander(a).parse(args);
+    Assert.assertEquals(a.params.size(), 1);
+//    Assert.assertEquals();
   }
 
   public void spaces() {