+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### Current
 
 2016-12-27
+* Fixed: Allow empty string (e.g. `java -jar jcommander-program.jar param1 ""`) as part of main parameter, #306 (@jeremysolarz)
 * Fixed: Default value for `@Parameter(help=true)` parameter is not displayed in output of `JCommander.usage()`, #305 (@jeremysolarz) 
 
 2016-12-16